bindings/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "roam-sdk"
version = "0.1.0"
description = "Python SDK for ROAM (Object Agent Mapper)"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "grpcio>=1.50.0",
    "protobuf>=4.21.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
]
dev = [
    "grpcio-tools>=1.50.0",
    "black",
    "isort",
    "mypy"
]

[tool.hatch.build.targets.wheel]
packages = ["roam_sdk"]

[dependency-groups]
dev = [
    "types-grpcio>=1.0.0.20251009",
    "types-protobuf>=6.32.1.20251210",
]
